feat(api-gateway): initialize API Gateway with routing and authentication middleware

+++ apps/blog-service/src/services/blog.service.ts
@@ -8,6 +8,14 @@ class BlogService {
         content: string,
         authorId: number,
     ) {
+        // check if this Blog already exists
+        const existingBlog = await db.orm.public.Blog.where({ slug }).first()
+        if(existingBlog) {
+            throw new ApiError(
+                409,
+                "Blog is already created"
+            )
+        }
         const blog = await db.orm.public.Blog.create({
             title,
             slug,
@@ -18,6 +26,12 @@ class BlogService {
     }
     async getAllBlogs () {
         const blogs = await db.orm.public.Blog.all();
+        if(!blogs) {
+            throw new ApiError(
+                404,
+                "Blogs not found"
+            )
+        }
         return blogs;
     }
     async getSingleBlog(id: number) {
@@ -32,15 +46,22 @@ class BlogService {
     }
     async updateBlog ( 
         id: number,
+        userId: number,
         title?: string,
         slug?: string,
-        content?: string
+        content?: string,
     ) {
         const existingBlog = await db.orm.public.Blog.where({ id }).first()
         if(!existingBlog) {
             throw new ApiError(
                 404,
                 "Blog not found"
+            )
+        }
+        if(existingBlog.authorId !== userId) {
+            throw new ApiError(
+                403,
+                "You are not allowed to update this blog"
             )
         }
         const blog = await db.orm.public.Blog.where({ id }).update({
@@ -50,12 +71,19 @@ class BlogService {
         });
         return blog;
     }
-    async deleteBlog (id: number) {
+    async deleteBlog (id: number, userId: number) {
         const existingBlog = await db.orm.public.Blog.where({ id }).first();
+
         if(!existingBlog) {
             throw new ApiError(
                 404,
                 "Blog not found"
+            )
+        }
+        if(existingBlog?.authorId !== userId) {
+            throw new ApiError(
+                403,
+                "You are not allowed to this action"
             )
         }
         await db.orm.public.Blog.where({ id }).delete()
